services/identity-svc/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic
pydantic-settings
cryptography
bcrypt
